fix: report only the delta of stable memory
Prior to this commit, `canbench` reported the size of the entire stable
memory as one if its measurements. While useful, a more useful metric
would be reporting the _delta_ in stable memory that was caused by the
benchmarked code, so that it's clearer to the user how their code is
causing changes to stable memory.