chore: introduce a type for measurements (#10)
Benchmark data were represented as a `BTreeMap` with string keys and
values. While this approach is flexible, it has some drawbacks:

1. When we introduce the option of profiling a specific part of the
code, there will be a lot of boilerplate code that's copied among
several places.
2. It's more difficult to ensure measurements contain all the needed
statistics.

This PR changes the representation of a benchmark measurement from a
`BTreeMap` to a `Measurement` type containing all of the required
measurements.