chore: nomenclature cleanup (#19)
The nomenclature in the `canbench` crate was not very clear. We exposed
the following APIs:

1. `canbench::bench`
2. `canbench::benchmark`
3. `canbench::profile`

These namings were historical. This commit
renames the APIs to the following:

1. `canbench::bench`: the macro used for declaring benchmarks
2. `canbench::bench_fn`: used for benchmarking a function
3. `canbench::bench_scope`: used for benchmarking the given scope.

Additionally, given that we've made changes to the API, I moved the
Rust-specific documentation into the crate's documentation. This has the
advantage that rust can ensure that the code samples remain correct, and
that the documentation is available on docs.rs whenever the library is
published.