feat: versioning support (#21)
Introduces versioning in `canbench_results.yml`. The file now includes
the version of `canbench` that generated the file, and `canbench` will
now refuse to run if the version in `canbench_results.yml` is newer than
itself.

This functionality is useful to resolve inconsistencies in a number of
situations:

1. Developers having different versions of `canbench`.
2. Github CI using a different version than what the developers are
using.

The Github CI example, once we open-source this repo, will be updated to
install the exact version of `canbench` that the developers are using to
provide consistent results.

NOTE: The benchmark CI jobs are failing because the change we're making
is not backward-compatible, and that's fine since we haven't released
`canbench` yet, so we'll ignore those failures this one time.