chore: Use RustLs instead of OpenSSL for reqwest (#47)
When using the default features of reqwest, it pulls in `native-tls`
which depends on `openssl`. Therefore, when such dependency exists in
the IC monorepo, it will fail the
[binary-libraries-test.sh](https://sourcegraph.com/github.com/dfinity/ic@4d851cbcf6b0379ae83411997e18358fbdba6e71/-/blob/publish/binaries/binary-libraries-test.sh).

When setting the default-features to false without adding
`rustls-tls-native-roots`, the tests fail because `drun` cannot be
downloaded with`https://`. Either `rustls-tls-native-roots` or
`rustls-tls-webpki-roots`seems to work and it doesn't seem to make much
difference here. Choosing `rustls-tls-webpki-roots` as it's the same as
the
[monorepo](https://sourcegraph.com/github.com/dfinity/ic@4d851cbcf6b0379ae83411997e18358fbdba6e71/-/blob/bazel/external_crates.bzl?L970)