perf: only install canister once when running benchmarks
Prior to this commit, PocketIC reinstalled the canister for every single
benchmark it runs. This was due to a limitation in `drun` (the runtime
used prior to PocketIC) and isn't unnecessary given that benchmarks are
all queries.

With this change, the canister is installed only once at the beginning,
resulting in a significant speed up when having lots of benchmarks.

Testing this change on the `stable-structures` crate:

Before this change:

```
canbench  18.17s user 2.81s system 12% cpu 2:46.37 total
```

With this change:

```
canbench  0.54s user 0.13s system 1% cpu 1:01.41 total
```

Benchmarks run almost 3x faster in this example.