feat: add flags to set a custom runtime.
## Problem
To enable running canbench smoothly in the `ic` repo, it is desirable to
able to run canbench with the runtime (PocketIC) that is in the same
commit as the PR. canbench didn't provide that flexibility, as it always
downloaded and verified that the runtime has a specific hash.

## Solution
Add two flags, one to set the path to the runtime, and another to skip
the hash verification step altogether.