fix: publishing crates with correct versions (#70)
## Problem
When publishing canbench previously, I had to manually change the local
dependencies to point to the versions in cargo (e.g. when publishing
`canbench-rs`, I need to change the dependency on `canbench-rs-macros`
from the local folder to the new version of `canbench-rs-macros` that I
just published). This process is messy and confusing, and is one of the
things that wouldn't be intuitive to anyone publishing canbench in the
future.

## Solution
Update Cargo.toml such that, locally, it uses a folder, and when
publishing, it uses a specific version. That way, publishing can be
seamless and requires no local modifications moving forward.

Co-authored-by: Dimitris Sarlis <sarlis.dim@gmail.com>