feat: Support custom noise threshold (#79)
# Why

For some projects, there are some small variances in the performance,
and from time to time people have to update the results for small
changes. For projects with more frequent commits, the chances of having
conflicts is greater, and the convenience of avoiding such conflicts
sometimes outweighs the benefits of detecting small performance changes.

# What

- Add `--custom-noise-threshold` CLI option
- Use the custom threshold if it exists
- Add tests

# Alternative Considered

One alternative is to add the argument to `Config` instead. There
doesn't seem to be a clear pros/cons for whether to add to CLI or
Config. For IC monorepo, it's easier to use CLI option because of bazel.