feat: Allow repeated scopes (#94)
This change allows for repeated scopes to be used in canbench
benchmarks. Previously the code would straight up refuse to have a
similarly named scope -- I think this is a mistake, there are very valid
scenarios ([example
case](https://github.com/dfinity/stable-structures/pull/296) from stable
structures) where someone might have a loop and they want to declare a
scope inside it to measure some operation.

People should already be careful not declaring the same names for
unrelated scopes, so I don't think a protection against it is worth it
(that's I guess the main reason for this restriction currently) and the
extra benefit you get to measure things inside loops (which can usually
be where heavier parts of processing happens) outweighs some risk that
someone re-uses the same name of scope.